requirements.txt
absl-py==2.2.2
aiofiles==23.2.1
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
altair==5.5.0
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.0.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
babel==2.17.0
boto3==1.38.23
botocore==1.38.23
cached_path==1.7.3
cachetools==5.5.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cn2an==0.5.22
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
crcmod==1.7
cryptography==44.0.2
cycler==0.12.1
dateparser==1.1.8
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
Distance==0.1.3
distro==1.9.0
dj-rest-auth==7.0.1
Django==4.2.21
django-allauth==65.8.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==3.2.3
django-redis==5.4.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-nested-routers==0.94.2
docopt==0.6.2
dotenv==0.9.9
eng-to-ipa==0.0.2
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.1.17
ffmpy==0.5.0
filelock==3.14.0
fonttools==4.58.0
fsspec==2025.5.1
fugashi==1.3.0
g2p-en==2.1.0
g2pkk==0.1.2
google-api-core==2.24.2
google-auth==2.40.2
google-cloud-core==2.4.3
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gradio==4.26.0
gradio_client==0.15.1
grpcio==1.71.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.1
gruut-lang-en==2.0.1
gruut-lang-es==2.0.1
gruut-lang-fr==2.0.2
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.20.3
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
inflect==7.0.0
ipykernel==6.29.5
ipython==8.18.1
jaconv==0.4.0
jamo==0.4.1
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.6
jiter==0.8.2
jmespath==0.10.0
joblib==1.5.1
jsonlines==1.2.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
langid==1.1.6
librosa==0.9.1
llvmlite==0.43.0
loguru==0.7.2
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.4
matplotlib-inline==0.1.6
mdurl==0.1.2
mecab-python3==1.0.9
model-index==0.1.11
mpmath==1.3.0
narwhals==1.40.0
nest-asyncio==1.6.0
networkx==2.8.8
nltk==3.9.1
num2words==0.5.12
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.2
openai==1.59.7
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.2
ordered-set==4.1.0
orjson==3.10.18
oss2==2.17.0
packaging==24.2
pandas==2.0.3
parso==0.8.3
pexpect==4.9.0
Pillow==9.5.0
plac==1.4.5
platformdirs==4.3.6
pooch==1.8.2
proces==0.1.7
prompt-toolkit==3.0.43
proto-plus==1.26.1
protobuf==6.31.0
psutil==6.1.1
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.0
pydantic_core==2.0.1
pydub==0.25.1
Pygments==2.17.2
PyJWT==2.9.0
pykakasi==2.2.1
pyparsing==3.2.3
pypinyin==0.50.0
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
python3-openid==3.2.0
pytz==2023.3
# pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
redis==6.1.0
referencing==0.36.2
regex==2024.11.6
requests==2.28.2
requests-oauthlib==2.0.0
resampy==0.4.3
rich==13.4.2
rpds-py==0.25.1
rsa==4.9.1
ruff==0.11.11
s3transfer==0.13.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
semantic-version==2.10.0
sentence-transformers==2.2.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.6.1
soundfile==0.13.1
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.46.2
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tokenizers==0.13.3
tomlkit==0.12.0
torch==2.7.0
torchaudio==2.7.0
tornado==6.4.2
tqdm==4.65.2
traitlets==5.14.0
transformers==4.27.4
txtsplit==1.0.0
typer==0.15.4
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.3.1
Unidecode==1.3.7
unidic==1.1.0
unidic-lite==1.0.8
urllib3==1.26.20
uvicorn==0.34.2
wasabi==0.10.1
wcwidth==0.2.12
websockets==11.0.3
Werkzeug==3.1.3
Wikipedia-API==0.7.3
win32_setctime==1.2.0
wrapt==1.17.2
zipp==3.21.0
gunicorn==21.2.0
